Move table of content markup generation into createGuide.js
Moves the generation of the table of content markup into the createGuide.js. This relies on the marked `Renderer` method `heading` which is called for each heading encountered in the Markdown source. For each level 3 or 4 heading, a nested array is populated which in structure matches the hierarchy of the table of content of the current section. This array is used to prepend the table of content markup to the final HTML.