Move markup generation back into Guide.svelte
Moves the generation of the table of contents markup back into the Guide.svelte file so that it is subject to Svelte’s own rendering process. This allows Svelte to attach its generated class names that are used to scope styles.

Also adds the `lang` attribute to the document when rendering a guide because the guides are available in multiple languages. Ideally, this would be done on the server side, but I don’t know how to do this with Sapper.