Add table of contents to subsections (#228)
* Add table of contents to subsections

Adds a generated, two-level deep table of contents to all subsections (i.e. to the documents found in the `guide/en/` directory). The elements in the table of contents are plain links pointing to fragment identifiers of their corresponding heading elements (i.e. the heading elements’ ID attributes). This is intended to improve the navigability of the documentation especially for sections like “Plugin Development” and “Big list of options” which both have many, many sub sections.

Moves the chain of `String.prototype.replace` calls sanitizing heading markup into its own function because it’s now also needed for the logic preparing the new table of content data. Also adds two new calls replacing the HTML entities `&lt;` and `&gt;` which appear in some of the headings (e.g. https://rollupjs.org/guide/en/#-p-plugin---plugin-plugin).

* Move table of content markup generation into createGuide.js

Moves the generation of the table of content markup into the createGuide.js. This relies on the marked `Renderer` method `heading` which is called for each heading encountered in the Markdown source. For each level 3 or 4 heading, a nested array is populated which in structure matches the hierarchy of the table of content of the current section. This array is used to prepend the table of content markup to the final HTML.

* Move markup generation back into Guide.svelte

Moves the generation of the table of contents markup back into the Guide.svelte file so that it is subject to Svelte’s own rendering process. This allows Svelte to attach its generated class names that are used to scope styles.

Also adds the `lang` attribute to the document when rendering a guide because the guides are available in multiple languages. Ideally, this would be done on the server side, but I don’t know how to do this with Sapper.

* Fixes rendering an empty <ul></ul> element

* Fixes heading anchors always pointing to English guide

The heading anchors always pointed to the English guide because the URL paths were hard-coded to `guide/en` in Guide.svelte and createGuide.js.